+++ indico/core/signals/event/registration.py
@@ -56,5 +56,13 @@ Called when resolving whether Indico should send tickets with e-mails
 or it will be handled by other module. The `sender` is the
 `RegistrationForm` object.
 
-Should return ``True`` or ``False``.
+If this signal returns ``True``, no ticket will be emailed on registration.
+""")
+
+is_ticket_blocked = _signals.signal('is-ticket-blocked', """
+Called when resolving whether Indico should let a registrant download
+their ticket.  The `sender` is the registrant's `Registration` object.
+
+If this signal returns ``True``, the user will not be able to download
+their ticket.
 """)